+++ src/components/Flow/components/Sidebar/styles.ts
