+++ src/components/Flow/todo.md
@@ -1,4 +1,3 @@
 * mudança do track que está sendo mostrado no gráfico
-* renomear tracks para flows ----> (OK)
 * adicionar nova track
 * permitir somente nome sem espaços e sem caracteres especiais
